==> server/ML_model/2_dataCleansing.ipynb <==
import pandas as pd
import datetime
from datetime import datetime

# 파일불러오기
tourList = pd.read_excel('티맵관광지목록_정리 완료.xlsx') 
tourList = tourList.iloc[:,[0,1,5]]
# df = pd.DataFrame(colums = ['','',''])

import datetime as dt

def leftjoin():
        date = datetime(2021,1,1)
        df = pd.DataFrame(columns = ['num','관광지명','검색건수','area_code','date','avgTemp','highTemp','lowTemp','weather','rainProb'])

        while True:       
                date_str = str(date)[:10]
                date4tmap = date_str.replace("-","")
                tmapSearch = pd.read_excel('./관광데이터/'+date4tmap+'.xlsx')        
                        # tmapSearch = pd.read_excel(date+'.xlsx')        
                join1 = pd.merge(left = tourList, right=tmapSearch, how = "left", on='관광지명')
                join1 = join1[['num','관광지명','검색건수','area_code']]
                join1 = join1.fillna(0)
                join1['datetime'] = date
                join1['date'] = date_str # 날짜형식 2021-10-10
                for i in range(0,12):
                        weatherForecast = pd.read_excel('./날씨 데이터/weatherAreaCode'+str(i)+'.xlsx')
                        join2 = pd.merge(left = join1, right=weatherForecast, how = "inner", on=['area_code','date'])
                        df = pd.concat([df,join2], ignore_index=True)
                date = date+dt.timedelta(days=1)
                if date==datetime(2022,3,1):
                        break
        df.to_excel('dataset2.xlsx')          
leftjoin()


# from dateutil.parser import parse

# date = dt.datetime(2021,1,1)       
# date_str = str(date)[:10]
# print(date_str.replace("-",""))


df = pd.read_excel('dataset2.xlsx')
df = df[['관광지명','검색건수','area_code','date','avgTemp','highTemp','lowTemp','weather','rainProb','datetime']]
df = df.dropna()
df['month'] = df['datetime'].dt.month_name()
# 월요일: 0, 일요일: 6
df['weekday'] = df['datetime'].dt.weekday
df.to_excel('dataset_clean.xlsx', index=False)
