==> server/ML_model/3_algorithmCompare.ipynb <==
# ! pip install tensorflow
# ! pip install xgboost
# ! pip install lightgbm


import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_validate

tourlist=pd.read_excel('dataset_clean.xlsx')
tourlist.info()

#범주형 데이터 라벨인코딩
from sklearn.preprocessing import LabelEncoder
encoder = LabelEncoder()
encoder.fit(tourlist['관광지명'])
tourlist['관광지명'] = encoder.transform(tourlist['관광지명'])
encoder.fit(tourlist['weather'])
tourlist['weather'] = encoder.transform(tourlist['weather'])

# 관광지명, 날씨, 평균온도, 최고온도, 최저온도, 강수확률
train = pd.concat([tourlist['관광지명'], tourlist['weather'], tourlist[['avgTemp','highTemp','lowTemp','rainProb']]], axis=1)
target = tourlist['검색건수']
train_input,test_input,train_target,test_target = train_test_split(
    train,target, test_size = 0.3)

# 라벨인코딩_랜덤포레스트
from sklearn.ensemble import RandomForestRegressor
model = RandomForestRegressor()
scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
print(np.mean(scores['train_score']),np.mean(scores['test_score']))

# 라벨인코딩_xgb
from xgboost import XGBRegressor
model = XGBRegressor(tree_method = 'hist')
scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
print(np.mean(scores['train_score']),np.mean(scores['test_score']))

# 라벨인코딩_lgbm
from lightgbm import LGBMRegressor
model = LGBMRegressor(tree_method = 'hist')
scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
print(np.mean(scores['train_score']),np.mean(scores['test_score']))

tourlist=pd.read_excel('dataset_clean.xlsx')

#범주형 데이터 원핫인코딩
from sklearn.preprocessing import LabelEncoder
tourlist_oh = pd.get_dummies(tourlist['관광지명'])
weather_oh = pd.get_dummies(tourlist['weather'])

# 관광지명, 날씨, 평균온도, 최고온도, 최저온도, 강수확률
train = pd.concat([tourlist_oh, weather_oh, tourlist[['avgTemp','highTemp','lowTemp','rainProb']]], axis=1)
target = tourlist['검색건수']
train_input,test_input,train_target,test_target = train_test_split(
    train,target, test_size = 0.3)

# 원핫인코딩_랜덤포레스트 >>자원활용 문제
# from sklearn.ensemble import RandomForestRegressor
# model = RandomForestRegressor()
# scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
# print(np.mean(scores['train_score']),np.mean(scores['test_score']))

# 원핫인코딩_xgb
from xgboost import XGBRegressor
model = XGBRegressor(tree_method = 'hist')
scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
print(np.mean(scores['train_score']),np.mean(scores['test_score']))

# 원핫인코딩_lgbm >>Do not support special JSON characters in feature name 에러
# from lightgbm import LGBMRegressor
# model = LGBMRegressor(tree_method = 'hist')
# scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
# print(np.mean(scores['train_score']),np.mean(scores['test_score']))

# 정규화
from sklearn.preprocessing import MinMaxScaler
tourlist4sc = tourlist[['avgTemp','highTemp','lowTemp','rainProb']]
scaler = MinMaxScaler()
scaler.fit(tourlist4sc)
tourlist_sc = scaler.transform(tourlist4sc)
tourlist_sc = pd.DataFrame(tourlist_sc, columns=['avgTemp','highTemp','lowTemp','rainProb'])

# 관광지명, 날씨, 평균온도, 최고온도, 최저온도, 강수확률
train = pd.concat([tourlist_oh, weather_oh, tourlist_sc], axis=1)
target = tourlist['검색건수']
train_input,test_input,train_target,test_target = train_test_split(
    train,target, test_size = 0.3)

# 원핫인코딩_정규화_xgb
from xgboost import XGBRegressor
model = XGBRegressor(tree_method = 'hist')
scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
print(np.mean(scores['train_score']),np.mean(scores['test_score']))

# 표준화
from sklearn.preprocessing import StandardScaler
tourlist4sc = tourlist[['avgTemp','highTemp','lowTemp','rainProb']]
scaler = StandardScaler()
scaler.fit(tourlist4sc)
tourlist_sc = scaler.transform(tourlist4sc)
tourlist_sc = pd.DataFrame(tourlist_sc, columns=['avgTemp','highTemp','lowTemp','rainProb'])

# 관광지명, 날씨, 평균온도, 최고온도, 최저온도, 강수확률
train = pd.concat([tourlist_oh, weather_oh, tourlist_sc], axis=1)
target = tourlist['검색건수']
train_input,test_input,train_target,test_target = train_test_split(
    train,target, test_size = 0.3)

# 원핫인코딩_표준화_xgb
from xgboost import XGBRegressor
model = XGBRegressor(tree_method = 'hist')
scores = cross_validate(model, train_input, train_target, return_train_score=True, n_jobs=-1)
print(np.mean(scores['train_score']),np.mean(scores['test_score']))
