==> server/ML_model/7_visitjejuScarp.ipynb <==
import requests
import time

def scrap(output_file_name,pagenum):
    url = 'https://api.visitjeju.net/api/contents/list?_siteId=jejuavj&locale=kr&device=pc&cate1cd=cate0000000002&tag=&sorting=reviewcnt+desc,+title_kr+asc&region1cd=&region2cd=&pageSize=12&page='+str(pagenum)

    headers = {
        'Accept':'application/json, text/plain, */*',
        'Accept-Encoding':'gzip, deflate, br',
        'Accept-Language':'ko-KR,ko;q=0.9,en-US;q=0.8,en;q=0.7',
        'Cache-Control':'no-cache',
        'Connection':'keep-alive',
        'Cookie':'_fbp=fb.1.1649665454601.3816748; _gcl_au=1.1.19800999.1649665455; iceJWT=SDP+eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJBbm9ueW1vdXMiLCJhdWQiOiJmNTI5ZDJjZS1mM2UxLTQwMWMtYTJmOC1iNjZhZGNkNjA5ZDUiLCJqdGkiOiJmNTI5ZDJjZS1mM2UxLTQwMWMtYTJmOC1iNjZhZGNkNjA5ZDUiLCJpc3MiOiJJLU9OIiwiaWF0IjoxNjQ5NzYyMTY3LCJleHAiOjE2NDk4NzAxNjd9.uZe-Tlq-rPsWqUiA5HR5V_GB-KtrZvL93nNXOeQQ_lQ1nkPX-MP64XsrablFkO7CRJZwXsb1TIsDjLr3KWVpKA; iceRefreshJWT=SDP+eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJBbm9ueW1vdXMiLCJhdWQiOiI1MzE0ODQxMi02NDdmLTQ0YWUtYjMzMS04MWQ2MGRjZDE2NzUiLCJpc3MiOiJJLU9OIiwianRpIjoiZmE5MjNmNmEtY2IwYi00ZTVlLWFiZDUtNDhmN2FhNjg3NzFkIiwiaWF0IjoxNjQ5NzYyMTY3LCJleHAiOjQxMDI0MTI0MDB9.t5zZAOMl_c6KE_flN-1WYmsMAlsgmngD1hT4RvO28hljyu-DnydRWa3kyib4giminxiLzSm8xXcJwWWulzkIjg; _gid=GA1.2.181648737.1649762169; Hm_lvt_f6d5e380b023c5b1e946de764e07a2a2=1649566628,1649665455,1649762169; _gat_gtag_UA_74784502_6=1; _ga=GA1.2.257577945.1649566627; Hm_lpvt_f6d5e380b023c5b1e946de764e07a2a2=1649770141; _ga_WR7WEFWP2T=GS1.1.1649765011.5.1.1649770148.0',
        'Host':'api.visitjeju.net',
        'Origin':'https://www.visitjeju.net',
        'Pragma':'no-cache',
        'Referer':'https://www.visitjeju.net/',
        'sec-ch-ua':'" Not A;Brand";v="99", "Chromium";v="100", "Google Chrome";v="100"',
        'sec-ch-ua-mobile':'?0',
        'sec-ch-ua-platform':"Windows",
        'Sec-Fetch-Dest':'empty',
        'Sec-Fetch-Mode':'cors',
        'Sec-Fetch-Site':'same-site',
        'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.88 Safari/537.36'
        }

   
    elements = requests.get(url, headers = headers).json()['items']

    results = []
    for element in elements:
        title = element['title']
        contentsid = element['contentsid']
        latitude = element['latitude']
        longitude = element['longitude']
        address = element['address']
        roadaddress = element['roadaddress']
        tag = element['tag']
        alltag = element['alltag']   
        try:
            introduction = element['introduction'].replace("\n", " ").replace("\t", " ").replace("\r", " ").replace("&nbsp;", " ").strip() 
        except:
            introduction = element['introduction']  
        try:
            sbst = element['sbst'].replace("\n", " ").replace("\t", " ").replace("\r", " ").replace("&nbsp;", " ").strip() 
        except:
            sbst = element['sbst']  
        imgpath = element['repPhoto']['photoid']['imgpath']
        thumbnailpath = element['repPhoto']['photoid']['thumbnailpath']

        results.append([
            title,
            contentsid,
            latitude,
            longitude,
            address,
            roadaddress,
            tag,
            alltag,
            introduction,
            sbst,
            imgpath,
            thumbnailpath
        ])
        fwrite_land(
            title,
            contentsid,
            latitude,
            longitude,
            address,
            roadaddress,
            tag,
            alltag,
            introduction,
            sbst,
            imgpath,
            thumbnailpath,
            output_file_name)
    return results


def fmake_file():
  output_file_name = 'visitjeju_info.csv'
  output_file = open(output_file_name, "w", encoding="utf-8")
  output_file.write("{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\n".format(
        'title',
        'contentsid',
        'latitude',
        'longitude',
        'address',
        'roadaddress',
        'tag',
        'alltag',
        'introduction',
        'sbst',
        'imgpath',
        'thumbnailpath',))
  output_file.close()
  return output_file_name

def fwrite_land(
        title,
        contentsid,
        latitude,
        longitude,
        address,
        roadaddress,
        tag,
        alltag,
        introduction,
        sbst,
        imgpath,
        thumbnailpath, 
        output_file_name):
    output_file = open(output_file_name, "a", encoding="utf-8")
    output_file.write("{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\n".format(
        title,
        contentsid,
        latitude,
        longitude,
        address,
        roadaddress,
        tag,
        alltag,
        introduction,
        sbst,
        imgpath,
        thumbnailpath,
    ))
    output_file.close()
    return

def fmain():
    output_file_name = fmake_file()
    for pagenum in (1,93):
        scrap(output_file_name,pagenum)
        print(pagenum)

fmain()

import pandas as pd
df_500 = pd.read_csv('visitJeju_areaCode.csv', sep='\t')
df_1000 = pd.read_csv('visitjeju_info.csv', sep='\t')

for i in zip(df_500.contentsid,df_500.title):
    print(i)
    df_1000.loc[df_1000.contentsid == i[0], 'title'] = i[1]


df_1000.to_excel('visitJejuList.xlsx')
