==> server/ML_model/8_concatTourlist.ipynb <==
import requests
import pandas as pd


tmap_list = pd.read_excel('관광지 주소 추가 전.xlsx')
visitJ = pd.read_csv('visitJeju_areaCode.csv', sep='\t', encoding='utf-8')

# df = pd.DataFrame(['tour','address','long','lat'])


df = pd.merge(tmap_list,visitJ,left_on='관광지명', right_on='title', how = 'left')
df.to_excel('티맵관광지목록_주소변경.xlsx', encoding='utf-8')




# headers = {"Authorization": "KakaoAK 5d6bbf017da7a866d48f509f80016557"}
# input_file = pd.read_excel('티맵네비게이션_관광지.xlsx')

# df = pd.DataFrame(columns=['areacode','tour','address','longtitde','latitude'])
# df['tour'] = input_file['관광지명']


# def adr():
#     for i in df['tour']:
#         # url= 'https://dapi.kakao.com/v2/local/keyword.json?query={1100고지 코스}.format('+i+')'
#         url= 'https://dapi.kakao.com/v2/local/search/keyword.json?query='+i
#         url= 'https://dapi.kakao.com/v2/local/search/address.json?query='+'서귀포잠수함'
#         result = requests.get(url, headers=headers).json()
#         print(result)
#         # address = result.get('address').get('address_name')
#         # long = result.get('address').get('x')
#         # lat = result.get('address').get('y')
#         # address_dong = result.get('road').get('region_3depth_name')
#         break
#     return 


# adr()

import pandas as pd
df_tmap = pd.read_excel('티맵관광지목록_정리 완료.xlsx')
df_visitj = pd.read_excel('visitJejuList.xlsx')

df_tmap.info()
df_visitj.info()

df = pd.merge(df_tmap, df_visitj, how='left', left_on='관광지명', right_on='title')
df

df.to_excel('제주도관광지정보.xlsx')

df.info()
